UI: Merge win-update and nix-update
