+++ UI/update/shared-update.hpp
@@ -3,6 +3,14 @@
 #include <QThread>
 #include <QString>
 
+#include <vector>
+#include <string>
+
+bool FetchAndVerifyFile(const char *name, const char *file, const char *url,
+			std::string *out,
+			const std::vector<std::string> &extraHeaders =
+				std::vector<std::string>());
+
 class WhatsNewInfoThread : public QThread {
 	Q_OBJECT
