+++ UI/update/win-update.hpp
@@ -27,29 +27,3 @@ public:
 	{
 	}
 };
-
-class WhatsNewInfoThread : public QThread {
-	Q_OBJECT
-
-	virtual void run() override;
-
-signals:
-	void Result(const QString &text);
-
-public:
-	inline WhatsNewInfoThread() {}
-};
-
-class WhatsNewBrowserInitThread : public QThread {
-	Q_OBJECT
-
-	QString url;
-
-	virtual void run() override;
-
-signals:
-	void Result(const QString &url);
-
-public:
-	inline WhatsNewBrowserInitThread(const QString &url_) : url(url_) {}
-};
