+++ UI/win-update/updater/helpers.cpp
@@ -1,0 +1,6 @@
+#include "helpers.hpp"
+
+void FreeProvider(HCRYPTPROV prov)
+{
+	CryptReleaseContext(prov, 0);
+}
